unique ids fam

+++ core/src/main/scala/com/github/andersonreyes/api/Body.scala
@@ -6,59 +6,59 @@ object Body {
   import com.github.andersonreyes.api.Config._
 
   @ConfiguredJsonCodec case class Echo(
-      msgId: Int,
+      msgId: Long,
       echo: String
   )
   @ConfiguredJsonCodec case class EchoOk(
-      msgId: Int,
+      msgId: Long,
       echo: String,
-      @JsonKey("in_reply_to") replyTo: Int
+      @JsonKey("in_reply_to") replyTo: Long
   )
 
   @ConfiguredJsonCodec case class Init(
-      msgId: Int,
+      msgId: Long,
       nodeId: String,
       nodeIds: List[String]
   )
   @ConfiguredJsonCodec case class InitOk(
-      @JsonKey("in_reply_to") replyTo: Int
+      @JsonKey("in_reply_to") replyTo: Long
   )
 
   @ConfiguredJsonCodec case class ErrorBody(
-      @JsonKey("in_reply_to") replyTo: Int,
-      code: Int,
+      @JsonKey("in_reply_to") replyTo: Long,
+      code: Long,
       text: String
   )
 
   @ConfiguredJsonCodec case class Generate(
-      msgId: Int
+      msgId: Long
   )
   @ConfiguredJsonCodec case class Topology(
       topology: Map[String, List[String]],
-      msgId: Int
+      msgId: Long
   )
 
   @ConfiguredJsonCodec case class TopologyOk(
-      @JsonKey("in_reply_to") replyTo: Int
+      @JsonKey("in_reply_to") replyTo: Long
   )
   @ConfiguredJsonCodec case class GenerateOk(
       id: String,
-      msgId: Int,
-      @JsonKey("in_reply_to") replyTo: Int
+      msgId: Long,
+      @JsonKey("in_reply_to") replyTo: Long
   )
 
   @ConfiguredJsonCodec case class Broadcast(
-      msgId: Int,
-      message: Int
+      msgId: Long,
+      message: Long
   )
 
   @ConfiguredJsonCodec case class BroadcastOk(
-      @JsonKey("in_reply_to") replyTo: Int
+      @JsonKey("in_reply_to") replyTo: Long
   )
 
-  @ConfiguredJsonCodec case class Read(msgId: Int)
+  @ConfiguredJsonCodec case class Read(msgId: Long)
   @ConfiguredJsonCodec case class ReadOk(
-      messages: List[Int],
-      @JsonKey("in_reply_to") replyTo: Int
+      messages: List[Long],
+      @JsonKey("in_reply_to") replyTo: Long
   )
 }
